# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RentUp API Documentation",
        "description": "OpenAPI 3.x documentation for the RentUp application.",
        "contact": {
            "email": "support@rentup.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://project.test/api/v1",
            "description": "Primary API Server"
        }
    ],
    "paths": {
        "/v1/devices": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Register a new device token",
                "description": "Stores a new Firebase/FCM device token for the authenticated user. Ability scope required: `devices:write`.",
                "operationId": "cf252444b25baacce38b78e8c061c51d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "fcm_token_12345ABCDEFG"
                                    },
                                    "platform": {
                                        "type": "string",
                                        "enum": [
                                            "ios",
                                            "android"
                                        ],
                                        "example": "android"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Device token stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device token registered."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden (Missing ability: devices:write)"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/devices/{device}": {
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Remove a device token",
                "description": "Deletes a specific device token for the authenticated user. Ability scope required: `devices:write`.",
                "operationId": "57b7453e23a7ceab499725a50b0fbb02",
                "parameters": [
                    {
                        "name": "device",
                        "in": "path",
                        "description": "ID of the device token record to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Device token deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden (Missing ability: devices:write)"
                    },
                    "404": {
                        "description": "Device token not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders",
                "description": "Fetches a paginated list of orders. Ability scope required: `orders:read`.",
                "operationId": "868a3eb6a92f63db88ce23b8a18a1198",
                "responses": {
                    "200": {
                        "description": "List of orders retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden (Missing ability: orders:read)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/orders/{order}/notify": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Manually trigger notification",
                "description": "Dispatches a job to send status change notifications. Ability scope required: `notify:send`.",
                "operationId": "3c816c0067d805dd1286e9121b0fc046",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID of the order to notify",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Notification job successfully enqueued",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notification job enqueued."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden (Missing ability: notify:send)"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a single order",
                "description": "Retrieves a specific order by ID. Requires 'orders:read' ability.",
                "operationId": "3fbf205ca31011e67fc11434d632ba98",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden (Missing 'orders:read' ability)"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [
                            "orders:read"
                        ]
                    }
                ]
            }
        },
        "/v1/orders/{order}/status": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status",
                "description": "Updates the status of an existing order. Requires 'orders:write' ability.",
                "operationId": "a02420b00f9a7a1d6f335d99b6ce9d06",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "processing",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden (Missing 'orders:write' ability)"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [
                            "orders:write"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Order": {
                "title": "Order Model",
                "description": "Order data model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 101
                    },
                    "user_id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 5
                    },
                    "code": {
                        "type": "string",
                        "example": "ORD-98765"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "shipped",
                            "delivered"
                        ],
                        "example": "pending"
                    },
                    "amount_decimal": {
                        "type": "number",
                        "format": "float",
                        "example": 49.99
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Global bearerAuth security scheme. Use your Sanctum API token.",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}